Deep Dive

+++ Fundamentals/18_promo/Samantha/Promote1.java
@@ -1,8 +1,8 @@
 public class Promote1{
     public static void main(String[] args) {
-        byte a = 5;
-        short b = 2;
-        int c = a +b;     
+        long a = 5;
+        float b = 2;
+        float c = a +b;     
         System.out.println(c);
     }
 }
